move robots.txt in sitemap
