+++ assessment/assessment-lib.js
@@ -56,43 +56,3 @@ export const createAssessment = async (userSite, userTitle) => {
     },
   };
 };
-
-export const getRobotsTxt = async (siteUrl) => {
-  const defaultReturnValue = {
-    sitemaps: null,
-    exists: false,
-    error: null,
-  };
-
-  const parseRobotsTxt = (robotsTxt) => {
-    try {
-      const regex = /Sitemap:\s*(https?:\/\/[^\s]+)/g;
-      let match;
-      const sitemaps = [];
-      // eslint-disable-next-line no-cond-assign
-      while ((match = regex.exec(robotsTxt)) !== null) {
-        sitemaps.push(match[1]);
-      }
-      return {
-        ...defaultReturnValue,
-        exists: true,
-        sitemaps: sitemaps.length > 0 ? sitemaps : null,
-      };
-    } catch (error) {
-      return { ...defaultReturnValue, ...{ exists: true, sitemaps: null, error } };
-    }
-  };
-
-  try {
-    const robotsResponse = await fetch(new URL('robots.txt', siteUrl).toString(), {
-      headers: { 'User-Agent': USER_AGENT },
-    });
-    if (robotsResponse.ok) {
-      const robotsTxt = await robotsResponse.text();
-      return parseRobotsTxt(robotsTxt);
-    }
-    return defaultReturnValue;
-  } catch (error) {
-    return { ...defaultReturnValue, error };
-  }
-};
